Implement debug logging of curl requests.
A couple of log statements were consolidated into the server.request method.

The log statements are now at debug level, and can basically be copy/pasted into a terminal to recreate the request.

There was also a sneaky issue resolved here where if there were any headers at all in a request with a body (and there always is the default User-Agent: Faraday), then no default Content-Type would be set.  This was apparent when looking though the new debug logs. I have changed this to just set the Content-Type to application/json if the Content-Type had specifically not been set.

The main reason for this change, however, is because I needed to know the difference between query parameters and the request body when debugging.